Refactored SearchDrawer to RightDrawer.
